26/04/2021 14:58

+++ spring-boot-web-application/spring-boot-first-application/src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -7,6 +7,8 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -15,13 +17,11 @@ import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.SessionAttributes;
 
 import com.in28minutes.springboot.web.model.Todo;
 import com.in28minutes.springboot.web.service.TodoService;
 
 @Controller
-@SessionAttributes("name")
 public class TodoController {
 	
 	@Autowired
@@ -34,21 +34,35 @@ public class TodoController {
 		binder.registerCustomEditor(Date.class, new CustomDateEditor(dateFormat, false));
 	}
 	
+	private String getLoginUserName(ModelMap model) {
+		Object principal = 
+				SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+		
+		if (principal instanceof UserDetails) {
+			return ((UserDetails) principal).getUsername();
+		}
+		
+		return principal.toString();
+	}
+	
 	@RequestMapping(value = "/list-todos", method = RequestMethod.GET)
     public String showTodosList(ModelMap model) {
-		String name = (String) model.get("name");
+		String name = getLoginUserName(model);
 		model.put("todos", service.retrieveTodos(name));
         return "list-todos";
 	}
-	
+
 	@RequestMapping(value = "/add-todo", method = RequestMethod.GET)
 	public String showAddTodoPage(ModelMap model){
-		model.addAttribute("todo", new Todo(0, (String) model.get("name"), "Default Desc", new Date(), false));
+		model.addAttribute("todo", new Todo(0, getLoginUserName(model), "Default Desc", new Date(), false));
 		return "todo";
 	}
 	
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id){
+		if (id == 1) {
+			throw new RuntimeException("Something went wrong");
+		}
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
@@ -65,7 +79,7 @@ public class TodoController {
 		if (result.hasErrors()) {
 			return "todo";
 		}
-		todo.setUser((String) model.get("name"));
+		todo.setUser(getLoginUserName(model));
 		
 		service.updateTodo(todo);
 		return "redirect:/list-todos";
@@ -77,7 +91,7 @@ public class TodoController {
 			return "todo";
 		}
 		
-		String name = (String) model.get("name");
+		String name = getLoginUserName(model);
 		service.addTodo(name, todo.getDesc(), todo.getTargetDate(), false);
 		return "redirect:/list-todos";
 	}
